requirements-dev.txt
# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-html>=3.1.0
pytest-xdist>=3.0.0

# Code Quality
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# Security
safety>=2.3.0
bandit>=1.7.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0

# Development Tools
pre-commit>=3.0.0
ipython>=8.0.0
jupyter>=1.0.0

# Coverage
coverage>=7.0.0
codecov>=2.1.0

# Configuration Management
PyYAML>=6.0.0
jsonschema>=4.0.0
